logic added to crud layers

+++ backend/src/features/learning/section/section-service.ts
@@ -5,6 +5,7 @@ import type { UpdateSectionDto } from "./dto/UpdateSectionDto.js";
 import { toDto } from "./section-mapper.js";
 import type { ISection } from "./section-model.js";
 import * as sectionRepository from "./section-repository.js";
+import { assertOwnerOfLearningPath, fetchLearningPathEntity } from "../learning-path/learningPath-service.js";
 
 
 export const oneSection = async (id: string) => {
@@ -14,7 +15,11 @@ export const oneSection = async (id: string) => {
 
 export const createSection = async (userId: string, learningPathId: string, createSectionDto: CreateSectionDto) => {
 
-    
+    const isOwner = await assertOwnerOfLearningPath(userId, learningPathId);
+    if (!isOwner) {
+        throw new HttpError(403, "You are not the owner of this learning path");
+    }
+
     const newSection: Omit<ISection, "_id" | "order"> = {
         name: createSectionDto.name,
         description: createSectionDto.description,
@@ -28,9 +33,9 @@ export const createSection = async (userId: string, learningPathId: string, crea
 
 export const updateSection = async (userId: string, sectionId: string, learningPathId: string, updateSectionDto: UpdateSectionDto) => {
 
-    const section = await fetchSectionEntity(sectionId);
-    if (section.learningPathId.toString() !== learningPathId) {
-        throw new HttpError(403, "You are not authorized to update this section");
+    const isOwner = await assertOwnerOfLearningPath(userId, learningPathId);
+    if (!isOwner) {
+        throw new HttpError(403, "You are not the owner of this learning path");
     }
 
     await sectionRepository.updateById(sectionId, updateSectionDto);
@@ -38,19 +43,31 @@ export const updateSection = async (userId: string, sectionId: string, learningP
 }
 
 export const deleteSection = async (userId: string, sectionId: string, learningPathId: string) => {
-   
+       
+    const isOwner = await assertOwnerOfLearningPath(userId, learningPathId);
+    if (!isOwner) {
+        throw new HttpError(403, "You are not the owner of this learning path");
+    }
+
     const section = await fetchSectionEntity(sectionId);
+
     if (section.learningPathId.toString() !== learningPathId) {
-        throw new HttpError(403, "You are not authorized to delete this section");
+        throw new HttpError(403, "This Section is not part of this Learning Path");
     }
 
     await sectionRepository.deleteById(sectionId);
 }
 
-const fetchSectionEntity = async (id: string) => {
+export const fetchSectionEntity = async (id: string): Promise<ISection> => {
     const section = await sectionRepository.findById(id);
     if (!section) {
         throw new HttpError(404, "Section not found");
     }
     return section;
+}
+
+export const assertOwnerOfSection = async (userId: string, sectionId: string): Promise<boolean> => {
+    const section = await fetchSectionEntity(sectionId);
+    const learningPath = await fetchLearningPathEntity(section.learningPathId.toString());
+    return learningPath.userId.toString() === userId;
 }
