+++ backend/src/features/learning/topic/topic-service.ts
@@ -5,6 +5,8 @@ import type { UpdateTopicDto } from "./dto/UpdateTopicDto.js";
 import { toDto } from "./topic-mapper.js";
 import { type ITopic } from "./topic-model.js";
 import * as topicRepository from "./topic-repository.js";
+import { assertOwnerOfSection, fetchSectionEntity } from "../section/section-service.js";
+import { fetchLearningPathEntity } from "../learning-path/learningPath-service.js";
 
 export const oneTopic = async (topicId: string) => {
     const topic = await fetchTopicEntity(topicId);
@@ -12,7 +14,12 @@ export const oneTopic = async (topicId: string) => {
     return toDto(topic);
 }
 
-export const createTopic = async (sectionId: string, createTopicDto: CreateTopicDto) => {
+export const createTopic = async (userId: string, sectionId: string, createTopicDto: CreateTopicDto) => {
+
+    const isOwner = await assertOwnerOfSection(userId, sectionId);
+    if (!isOwner) {
+        throw new HttpError(403, "You are not the owner of this section");
+    }
 
     const newTopic: Omit<ITopic, "_id"> = {
         name: createTopicDto.name,
@@ -25,12 +32,23 @@ export const createTopic = async (sectionId: string, createTopicDto: CreateTopic
     return topic;
 }
 
-export const updateTopic = async (topicId: string, updateTopicDto: UpdateTopicDto) => {
+export const updateTopic = async (userId: string, topicId: string, updateTopicDto: UpdateTopicDto) => {
+    const isOwner = await assertOwnerOfTopic(userId, topicId);
+    if (!isOwner) {
+        throw new HttpError(403, "You are not the owner of this topic");
+    }
+
     const topic = await topicRepository.updateById(topicId, updateTopicDto);
     return topic;
 }
 
-export const deleteTopic = async (topicId: string) => {
+export const deleteTopic = async (userId: string, topicId: string) => {
+
+    const isOwner = await assertOwnerOfTopic(userId, topicId);
+    if (!isOwner) {
+        throw new HttpError(403, "You are not the owner of this topic");
+    }
+
     const topic = await topicRepository.deleteById(topicId);
     return topic;
 }
@@ -43,3 +61,15 @@ const fetchTopicEntity = async (topicId: string) => {
     }
     return topic;
 }
+
+const assertOwnerOfTopic = async (userId: string, topicId: string): Promise<boolean> => {
+    const topic = await fetchTopicEntity(topicId);
+    const section = await fetchSectionEntity(topic.sectionId.toString());
+    const learningPath = await fetchLearningPathEntity(section.learningPathId.toString());
+
+    if (learningPath.userId.toString() !== userId) {
+        return false;
+    }
+
+    return true;
+};
